+++ tests/unit/util/test_configuration.py
@@ -711,20 +711,19 @@ output:
         assert config._scheduler.jobs, "not cancelled job"
         assert config.config_refresh_interval == 60, "should be None"
 
-    def test_reload_does_not_reload_but_logs_info_on_same_config(self, config_path, caplog):
+    def test_reload_always_reloads(self, config_path, caplog):
         caplog.set_level("INFO")
         config = Configuration.from_sources([str(config_path)])
         config._metrics.number_of_config_refreshes = 0
         config._metrics.number_of_config_refresh_failures = 0
-        # set process_count to a different value and save it assert that it is not reloaded
         config.process_count = 99
         config_path.write_text(config.as_yaml())
         config.process_count = 2
         config.reload()
-        assert "Configuration version didn't change." in caplog.text
-        assert config.process_count == 2
-        assert config._metrics.number_of_config_refreshes == 0, "no config refresh"
-        assert config._metrics.number_of_config_refresh_failures == 0, "no config refresh failure"
+        assert "Successfully reloaded configuration" in caplog.text
+        assert config.process_count == 99
+        assert config._metrics.number_of_config_refreshes == 1, "config refresh"
+        assert config._metrics.number_of_config_refresh_failures == 0, "config refresh failure"
 
     def test_reload_logs_error_on_invalid_config(self, config_path, caplog):
         config = Configuration.from_sources([str(config_path)])
@@ -910,16 +909,25 @@ output:
         assert config.version == "1", "version should be 1"
         config_path.unlink()  # causes FileNotFoundError
         config.reload()
-        assert "Failed to load configuration" in caplog.text
+        assert caplog.text.count("Failed to load configuration") == 1
         assert config.config_refresh_interval == 5, "set minimum refresh interval to 5"
         config.config_refresh_interval = 8  # to write original config
         config_path.write_text(config.as_yaml())
         config.config_refresh_interval = 5
         config.reload()
-        assert "Config refresh recovered from failing source" in caplog.text
-        assert "Config refresh interval is set to: 8 seconds" in caplog.text
+        assert caplog.text.count("Failed to load configuration") == 1
+        assert caplog.text.count("Config refresh recovered from failing source") == 1
+        assert caplog.text.count("Config refresh interval is set to: 8 seconds") == 2
         assert config.config_refresh_interval == 8, "refresh interval should be reset to origin"
-        assert config._metrics.number_of_config_refreshes == 0, "no refresh after recovering"
+        assert config._metrics.number_of_config_refreshes == 1, "refresh after recovering"
+        assert config._metrics.number_of_config_refresh_failures == 1, "config refresh failure"
+
+        config.reload()
+        assert caplog.text.count("Failed to load configuration") == 1
+        assert caplog.text.count("Config refresh recovered from failing source") == 1
+        assert caplog.text.count("Config refresh interval is set to: 8 seconds") == 4
+        assert config.config_refresh_interval == 8, "refresh interval should be reset to origin"
+        assert config._metrics.number_of_config_refreshes == 2, "refresh after recovering"
         assert config._metrics.number_of_config_refresh_failures == 1, "config refresh failure"
 
     def test_as_dict_returns_config(self):
@@ -1029,7 +1037,7 @@ output:
             ),
             (
                 "document is not a valid json or yaml",
-                "logprep.util.getter.FileGetter.get_yaml",
+                "logprep.util.getter.FileGetter.get_dict",
                 ScannerError,
                 "Invalid yaml or json file:",
             ),
@@ -1387,13 +1395,13 @@ output:
             )
         )
 
-    def test_log_config_refresh_interval_only_if_it_change(self, config_path, caplog):
+    def test_always_log_config_refresh_interval(self, config_path, caplog):
         caplog.set_level("INFO")
         config = Configuration.from_sources([str(config_path)])
         config.config_refresh_interval = 10
         config.reload()
-        assert "Configuration version didn't change." in caplog.text
-        assert "Config refresh interval is set to:" not in caplog.text
+        assert "Successfully reloaded configuration" in caplog.text
+        assert "Config refresh interval is set to:" in caplog.text
 
     def test_config_refresh_interval_cant_be_set_to_none(self, config_path, caplog):
         caplog.set_level("INFO")
@@ -1402,9 +1410,60 @@ output:
         config_path.write_text(config.as_yaml())
         config.config_refresh_interval = 10
         config.reload()
-        assert "Configuration version didn't change." in caplog.text
-        assert "Config refresh interval is set to:" not in caplog.text
+        assert "Successfully reloaded configuration" in caplog.text
+        assert "Config refresh interval is set to: 10 seconds" in caplog.text
         assert config.config_refresh_interval == 10, "should not be changed to None"
+
+    @responses.activate
+    def test_log_config_refresh_interval_change_independent_of_config_changes(
+        self, config_path, caplog
+    ):
+        caplog.set_level("INFO")
+        config = Configuration.from_sources([str(config_path)])
+
+        config.config_refresh_interval = 10
+
+        # Set to variable value
+        for processor in config.pipeline:
+            if "pseudonymizer" in processor:
+                processor["pseudonymizer"]["regex_mapping"] = "http://127.0.0.1:8000"
+                break
+
+        responses.get(
+            "http://127.0.0.1:8000",
+            json={
+                "RE_WHOLE_FIELD": ".",
+                "RE_DOMAIN_BACKSLASH_USERNAME": ".",
+                "RE_IP4_COLON_PORT": ".",
+                "RE_ALL_NO_CAP": ".",
+            },
+        )
+
+        config_path.write_text(config.as_yaml())
+        config.reload()  # Reload with newly added variable value
+        assert "Successfully reloaded configuration" in caplog.text
+
+        caplog.clear()
+        config.reload()  # Reload with unchanged variable value
+        assert "Successfully reloaded configuration" in caplog.text
+
+        responses.get(
+            "http://127.0.0.1:8000",
+            json={
+                "RE_WHOLE_FIELD": ".CHANGED",
+                "RE_DOMAIN_BACKSLASH_USERNAME": ".",
+                "RE_IP4_COLON_PORT": ".",
+                "RE_ALL_NO_CAP": ".",
+            },
+        )
+
+        config.reload()  # Reload with changed variable value
+        assert "Successfully reloaded configuration" in caplog.text
+
+        config = Configuration.from_sources([str(config_path)])  # Load with existing variable value
+        caplog.clear()
+        config.reload()  # Reload with unchanged variable value
+        assert "Successfully reloaded configuration" in caplog.text
 
 
 class TestInvalidConfigurationErrors: